updated from giac_1.2.0-15

+++ src/gen.h
@@ -733,7 +733,7 @@ namespace giac {
   class vectpoly:public std::vector<polynome> {
   public:
     vectpoly():std::vector<polynome>::vector() {};
-    vectpoly(int i,const polynome & p):std::vector<polynome>::vector(i,p) {};
+    vectpoly(size_t i,const polynome & p):std::vector<polynome>::vector(i,p) {};
     void dbgprint(){  
 #ifndef NSPIRE
       CERR << *this << std::endl;
